Change Fiji coordinates
Fiji were split and drawn at the opposite edges of the map. This caused their bounding box to extend for the whole map, in turn causing issues to click validation (a click in the Atlantic Ocean could be seen as intersecting Fiji islands).
Now the problem has been solved by removing the coordinates of Fiji overflowing on the left-most side of the map